+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/BroadcastChannel.kt
@@ -11,29 +11,11 @@ fun <T> BroadcastChannel<T>.subscribeChecking(
     },
     by: suspend (T) -> Boolean
 ): ReceiveChannel<T> {
-    return openSubscription().also {
-        launch {
-            while (isActive && !it.isClosedForReceive) {
-                try {
-                    val received = it.receive()
-
-                    launch {
-                        try {
-                            if (!by(received)) {
-                                it.cancel()
-                            }
-                        } catch (e: Throwable) {
-                            if (!throwableHandler(e)) {
-                                it.cancel()
-                            }
-                        }
-                    }
-                } catch (e: CancellationException) {
-                    break
-                }
-            }
-            it.cancel()
-        }
+    return openSubscription().apply {
+        subscribeChecking(
+            throwableHandler,
+            by
+        )
     }
 }
 
@@ -44,43 +26,85 @@ fun <T> BroadcastChannel<T>.subscribe(
     },
     by: suspend (T) -> Unit
 ): ReceiveChannel<T> {
+    return openSubscription().apply {
+        subscribeChecking(throwableHandler) {
+            by(it)
+            true
+        }
+    }
+}
+
+fun <T> ReceiveChannel<T>.subscribeChecking(
+    throwableHandler: (Throwable) -> Boolean = {
+        it.printStackTrace()
+        true
+    },
+    by: suspend (T) -> Boolean
+) {
+    val channel = this
+    launch {
+        while (isActive && !channel.isClosedForReceive) {
+            try {
+                val received = channel.receive()
+
+                launch {
+                    try {
+                        if (!by(received)) {
+                            channel.cancel()
+                        }
+                    } catch (e: Throwable) {
+                        if (!throwableHandler(e)) {
+                            channel.cancel()
+                        }
+                    }
+                }
+            } catch (e: CancellationException) {
+                break
+            }
+        }
+        channel.cancel()
+    }
+}
+
+fun <T> ReceiveChannel<T>.subscribe(
+    throwableHandler: (Throwable) -> Boolean = {
+        it.printStackTrace()
+        true
+    },
+    by: suspend (T) -> Unit
+) {
     return subscribeChecking(throwableHandler) {
         by(it)
         true
     }
 }
 
-fun <T> ReceiveChannel<T>.debounce(delayMs: Long): BroadcastChannel<T> {
-    val channel = BroadcastChannel<T>(Channel.CONFLATED)
-    var lastReceived: Pair<Long, T>? = null
-    var job: Job? = null
-    launch {
-        while (isActive && !isClosedForReceive) {
-            val received = receive()
 
-            lastReceived = Pair(System.currentTimeMillis() + delayMs, received)
+fun <T> BroadcastChannel<T>.debounce(
+    delay: Long,
+    timeUnit: TimeUnit = TimeUnit.MILLISECONDS
+): BroadcastChannel<T> {
+    return openSubscription().debounce(
+        delay,
+        timeUnit
+    )
+}
 
-            job ?:let {
-                job = launch {
-                    try {
-                        var now = System.currentTimeMillis()
-                        while (isActive && lastReceived?.first ?: now >= now) {
-                            delay((lastReceived ?.first ?: now) - now, TimeUnit.MILLISECONDS)
-                            now = System.currentTimeMillis()
-                        }
 
-                        lastReceived?.second?.also {
-                            channel.send(it)
-                        }
-                    } catch (e: Exception) {
-                        e.printStackTrace()
-                    } finally {
-                        job = null
-                    }
-                }
+fun <T> ReceiveChannel<T>.debounce(
+    delay: Long,
+    timeUnit: TimeUnit = TimeUnit.MILLISECONDS
+): BroadcastChannel<T> {
+    return BroadcastChannel<T>(Channel.CONFLATED).also {
+        outBroadcast ->
+        var lastReceived: Job? = null
+        subscribe {
+            lastReceived ?.cancel()
+            lastReceived = launch {
+                delay(delay, timeUnit)
+
+                outBroadcast.send(it)
             }
         }
-        cancel()
     }
-    return channel
 }